simulation/crow-report-web/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
python-dotenv
slowapi
Pillow
imagehash
GPSPhoto